requirements.txt
boto3>=1.34.0
pandas>=2.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
flask>=3.0.0
plotly>=5.18.0
